* commit 'f0c80a2e870f6517709220485bb4803e2a6f4f89':
  Change subId to int from long

+++ src/java/com/android/internal/telephony/Phone.java
@@ -1824,7 +1824,7 @@ public interface Phone {
     /*
      * Returns the subscription id.
      */
-    public long getSubId();
+    public int getSubId();
 
     /*
      * Returns the phone id.
